app/pyproject.toml
[tool.poetry]
name = "fast-sentiment"
version = "1.0.0"
description = "A high-performance sentiment analysis API using DistilBERT and FastAPI"
authors = ["Anna Sba <annasba@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
transformers = "^4.50.2"
torch = "^2.1.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.9"}
fastapi = {extras = ["all"], version = "^0.115.0"}
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.6.3"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.1.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"